Refactor Init and Train structs. Add hmm_test_utils.hpp
