+++ src/mlpack/tests/CMakeLists.txt
@@ -149,6 +149,7 @@ add_executable(mlpack_test
   main_tests/hmm_train_test.cpp
   main_tests/hmm_loglik_test.cpp
   main_tests/hmm_generate_test.cpp
+  main_tests/hmm_test_utils.hpp
 )
 
 # Link dependencies of test executable.
